% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pathways.R
\name{compare_pathways}
\alias{compare_pathways}
\title{Compare epiCHAOS scores across selected gene sets or biological pathways}
\usage{
compare_pathways(
  counts,
  set.names,
  min.row = 10,
  genome = "hg19",
  tss.span = c(1500, 500)
)
}
\arguments{
\item{counts}{a binarised sparse matrix containing regions/peaks-by-cells scATAC or other single-cell epigenomics dataset. Rownames should contain the genomic positions of peaks or regions in the format: "chrX:start-end"}

\item{set.names}{a vector of gene sets from msigdbr on which to compare epiCHAOS scores}

\item{min.row}{the minimum number of rows/peaks overlapping with promoters of selected gene sets. If the number of rows overlapping with promoters of a selected gene set is too few, epiCHAOS scores are not computed for that gene set.}

\item{genome}{the genome build e.g. "hg19", "hg38" on which to define promoter-overlapping regions. This defaults to hg19}

\item{tss.span}{a numeric vector containing the number of bp upstream and downstream of the TSS by which promoters should be defined. This defaults to c(1500, 500) i.e. 1500bp upstream and 500bp downstream of the TSS.}
}
\value{
a dataframe containing epiCHAOS scores for each selected gene set
}
\description{
Compare epiCHAOS scores across selected gene sets or biological pathways
}
\examples{
library(magrittr)
counts <- hemato_atac$counts
rownames(counts) <- paste0(hemato_atac$ranges)

# load hallmark gene ontologies for human from msigdb
gene_sets = msigdbr::msigdbr(species = "Homo sapiens")
gene_sets <- gene_sets[gene_sets$gs_cat=="H",]
set.names <- gene_sets$gs_name \%>\% unique()
het <- compare_pathways(counts=counts, set.names=set.names, min.row=5,
genome="hg19", tss.span=c(1500, 500))
}
