% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_epiCHAOS.R
\name{compute.eITH.cancer}
\alias{compute.eITH.cancer}
\title{Compute epiCHAOS scores with per-chromosome count correction, for application to cancer datasets where large copy number alterations may result in differences in total counts per chromosome}
\usage{
compute.eITH.cancer(x)
}
\arguments{
\item{x}{A list of binarised single-cell matrices on which to compute epiCHAOS scores with per-chromosome counts adjustment. The rownames of each matrix should indicate the chromosome, start and end coordinates separated by ":", "-" or "_".}
}
\value{
A dataframe containing epiCHAOS scores.
}
\description{
Compute epiCHAOS scores with per-chromosome count correction, for application to cancer datasets where large copy number alterations may result in differences in total counts per chromosome
}
\examples{
heterogeneity <- compute.eITH.cancer(matrices)
}
