% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_epiCHAOS.R
\name{compute.eITH}
\alias{compute.eITH}
\title{Compute epiCHAOS scores on a list of matrices}
\usage{
compute.eITH(x)
}
\arguments{
\item{x}{A list of binarised scATAC or other single cell matrices with equal rows.}
}
\value{
A dataframe containing epiCHAOS scores.
}
\description{
Compute epiCHAOS scores on a list of matrices
}
\examples{
heterogeneity <- compute.eITH(x)
}
