% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transCHAOS.R
\name{compute.tITH}
\alias{compute.tITH}
\title{Calculate transcriptional heterogeneity in single-cell RNA-seq data}
\usage{
compute.tITH(x)
}
\arguments{
\item{x}{a list of normalised scRNA-seq matrices, each matrix corresponding to a cell group/cluster of interest}
}
\value{
a dataframe containing transcriptional heterogeneity scores
}
\description{
Calculate transcriptional heterogeneity in single-cell RNA-seq data
}
\examples{
gex <- hemato_rna$data.scale
meta <- hemato_rna$meta
matrices <- create.group.matrices(gex, meta, colname="Clusters", binarise=FALSE)
heterogeneity <- compute.tITH(matrices)
}
