% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{compute_diff_eICH}
\alias{compute_diff_eICH}
\title{Function to compute differential epigenetic heterogeneity between two groups using epiCHAOS scores.}
\usage{
compute_diff_eICH(group1, group2, region.type, niter = 1000)
}
\arguments{
\item{group1}{A binarised peaks-by-cells matrix (or other single cell epigenomics matrix) for the first comparison group.}

\item{group2}{A binarised peaks-by-cells matrix (or other single cell epigenomics matrix) for the second comparison group.}

\item{region.type}{The name of the selected region type e.g. the TFBS.}

\item{niter}{A numeric value indicating the number of permutations. Defaults to 1000.}
}
\value{
A list containing (i) "gg": a density plot comparing the difference in epiCHAOS score between the two tested groups, compared to the distribution from permuted groups,
(ii) "result.mix": epiCHAOS scores from the permuted mixtures, and
(iii) "result.test": epiCHAOS scores from the test groups.
}
\description{
Function to compute differential epigenetic heterogeneity between two groups using epiCHAOS scores.
}
