% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_epiCHAOS.R
\name{create.group.matrices}
\alias{create.group.matrices}
\title{Create group matrices}
\usage{
create.group.matrices(counts, meta, colname, n = 100, m = 20, index = NULL)
}
\arguments{
\item{counts}{A counts matrix representing e.g. a peaks-by-cells matrix in the case of scATAC data. The counts matrix will be binarised if not already.}

\item{meta}{Metadata including the column on which the data should be grouped, e.g. cluster, cell type.}

\item{colname}{The name of the column in the metadata on which to group the data e.g. "cluster" or "celltype". If not specified this defaults the the first column in the provided metadata.}

\item{n}{The number of cells to subset for each group/cluster. This defaults to 100 cells.}

\item{m}{The minimum number of cells per group/cluster. This defaults to 20 cells. If fewer than m cells are found in a group/cluster, an epiCHAOS score is not computed for that group.}

\item{index}{The rows on which to subset the counts matrix. If not provided the whole counts matrix will be used by default. Otherwise "index" may be specified as either a vector of numerical indices or a vector or names corresponding to the rownames of interest in "counts".}
}
\value{
A list of binarised peaks/regions-by-cells matrices, one for each cluster/group of interest.
}
\description{
Create group matrices
}
\examples{
matrices <- compute.group.matrices(counts, metadata, "seurat_clusters", m=50)
}
