% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encode_tfbs}
\alias{encode_tfbs}
\title{Example regions set}
\format{
\subsection{\code{encode_tfbs}}{

A list of two components:
(i) ranges: a GRangesList object containing a GRanges object corresponding to each narrowPeak file from ENCODE TFBS
(ii) anno: a dataframe with four columns; filename (corresponding to names of (i)), celltype, treatment and antibody
}
}
\source{
\url{http://cloud.databio.org/regiondb/} "encode_tfbs"
}
\usage{
encode_tfbs
}
\description{
Genomic coordinates in GRanges format for a range of chromatin/transcription factor binding sites from the ENCODE TFBS project, subset for 10 regions to use as example
}
\keyword{datasets}
