% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_epiCHAOS.R
\name{epiCHAOS}
\alias{epiCHAOS}
\title{Compute epiCHAOS scores given a counts matrix and metadata}
\usage{
epiCHAOS(
  counts,
  meta,
  colname = colnames(meta)[1],
  n = 100,
  index = NULL,
  plot = F,
  cancer = F,
  subsample = 1
)
}
\arguments{
\item{counts}{A counts matrix representing e.g. a peaks-bycells matrix in the case of scATAC data. The counts matrix will be binarised if not already.}

\item{meta}{A dataframe containing metadata including at least the column on which the data should be grouped, e.g. cluster or cell type.}

\item{colname}{The name of the column in "meta" on which to group the data e.g. "cluster" or "celltype". If not specified this defaults the the first column in "meta".}

\item{n}{The number of cells to subset for each group/cluster. This defaults to 100.}

\item{index}{The rows on which to subset the counts matrix. If not provided the whole counts matrix will be used by default. Otherwise "index" may be specified as either a vector of numerical indices or a vector or names corresponding to the rownames of interest in "counts".}

\item{plot}{Defaults to FALSE. If TRUE, a bar plot will be returned as well as the epiCHAOS scores.}

\item{cancer}{Defaults to FALSE. If TRUE, the count correction step will be performed per-chromosome in order to take into account differences in coverage arising from large copy number alterations}

\item{subsample}{The number of times cells should be subsampled as "pseudoreplicates" for between-group comparisons. This defaults to 1. If > 1, "n" cells are subsampled from each group the specified number of times.}
}
\value{
A dataframe containing epiCHAOS scores for each group.
}
\description{
Compute epiCHAOS scores given a counts matrix and metadata
}
\examples{
heterogeneity <- epiCHAOS(counts, metadata, "seurat_cluster", subsample=10)
}
