% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hemato_rna}
\alias{hemato_rna}
\title{scRNA-seq Hematopoiesis data}
\format{
\subsection{\code{hemato_rna}}{

A list of three components:
(i) data.norm: a sparse matrix containing normalised scRNA-seq counts
(ii) data.scale: a sparse matrix containing scaled normalised scRNA-seq counts
(iii) meta: cell metadata including cluster and cell-type annotation corresponding to colnames of counts
For testing/example purposes the data were subset for progenitor cells, B, T and NK cells.
}
}
\source{
\url{https://jeffgranja.s3.amazonaws.com/MPAL-10x/Supplementary_Data/Healthy-Data/scRNA-Healthy-Hematopoiesis-191120.rds} "hemato_rna"
}
\usage{
hemato_rna
}
\description{
single-cell RNA-seq data from human hematopoiesis (Granja et al. 2019), subset for bone-marrow-derived cells
}
\keyword{datasets}
