% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_regions.R, R/differential.R,
%   R/epiCHAOS.R, R/transCHAOS.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{subsetByOverlaps}
\alias{GRanges}
\alias{aes}
\alias{lm}
\alias{jaccard}
\alias{str_split}
\alias{dist}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GenomicRanges}{\code{\link[GenomicRanges:GRanges-class]{GRanges}}}

  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{aes}}}

  \item{IRanges}{\code{\link[IRanges:findOverlaps-methods]{subsetByOverlaps}}}

  \item{jaccard}{\code{\link[jaccard]{jaccard}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}

  \item{stats}{\code{\link[stats]{dist}}, \code{\link[stats]{lm}}}

  \item{stringr}{\code{\link[stringr]{str_split}}}
}}

