% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_epiCHAOS.R
\name{compute.eITH.cancer}
\alias{compute.eITH.cancer}
\title{Compute epiCHAOS scores with per-chromosome count correction, for application to cancer datasets where large copy number alterations may result in differences in total counts per chromosome}
\usage{
compute.eITH.cancer(x)
}
\arguments{
\item{x}{A list of binarised single-cell matrices on which to compute epiCHAOS scores with per-chromosome counts adjustment. The rownames of each matrix should indicate the chromosome, start and end coordinates separated by ":", "-" or "_".}
}
\value{
A dataframe containing epiCHAOS scores.
}
\description{
Compute epiCHAOS scores with per-chromosome count correction, for application to cancer datasets where large copy number alterations may result in differences in total counts per chromosome
}
\examples{
library(magrittr)
m1 <- sample(c(0,1), 10*10, prob=c(0.8,0.2), replace = TRUE) \%>\% matrix(10,10)
m2 <- sample(c(0,1), 10*10, prob=c(0.8,0.2), replace = TRUE) \%>\% matrix(10,10)
rownames <- paste0(c("chr1", "chr2"), ":", seq(1000,100000, 10000), "-",
seq(2000,101000, 10000))
rownames(m1) <- rownames(m2) <- rownames
colnames(m1) <- colnames(m2) <- paste0("cell", 1:10)
x <- list(m1=m1, m2=m2)
heterogeneity <- compute.eITH.cancer(x)
}
