% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hemato_rna}
\alias{hemato_rna}
\title{Example scRNA-seq data}
\format{
\subsection{\code{hemato_rna}}{

A list of two components:
(i) data.scale: a sparse matrix containing scaled normalised scRNA-seq counts for subset of 100 HSCs, B-cells and T-cells
(ii) meta: cell metadata including cluster and cell-type annotation corresponding to colnames of counts
}
}
\source{
\url{https://jeffgranja.s3.amazonaws.com/MPAL-10x/Supplementary_Data/Healthy-Data/scRNA-Healthy-Hematopoiesis-191120.rds} "hemato_rna"
}
\usage{
hemato_rna
}
\description{
single-cell RNA-seq data from human hematopoiesis (Granja et al. 2019), subset for bone-marrow-derived cells, subset for 5000 genes to use for example purposes
}
\keyword{datasets}
