% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_epiCHAOS.R, R/differential.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{aes}
\alias{lm}
\alias{jaccard}
\alias{str_split}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{aes}}}

  \item{jaccard}{\code{\link[jaccard]{jaccard}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{stats}{\code{\link[stats]{lm}}}

  \item{stringr}{\code{\link[stringr]{str_split}}}
}}

